# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DriverOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "approved_accounts": {
            "items": {
              "type": "string"
            },
            "title": "Approved Accounts",
            "type": "array"
          },
          "assigned_trips": {
            "default": 0,
            "title": "Assigned Trips",
            "type": "integer"
          },
          "channel_today": {
            "title": "Channel Today",
            "type": "string"
          },
          "driver_id": {
            "format": "uuid",
            "title": "Driver Id",
            "type": "string"
          },
          "end_lat": {
            "title": "End Lat",
            "type": "number"
          },
          "end_lng": {
            "title": "End Lng",
            "type": "number"
          },
          "max_hours": {
            "title": "Max Hours",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "odometer_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Odometer Km"
          },
          "service_date": {
            "format": "date",
            "title": "Service Date",
            "type": "string"
          },
          "shift_end_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift End At"
          },
          "shift_start_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Start At"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "title": "Skills",
            "type": "array"
          },
          "soc_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soc Pct"
          },
          "soc_reported_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soc Reported At"
          },
          "start_lat": {
            "title": "Start Lat",
            "type": "number"
          },
          "start_lng": {
            "title": "Start Lng",
            "type": "number"
          },
          "vehicle_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Id"
          },
          "vehicle_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Model"
          },
          "vehicle_registration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Registration"
          },
          "vehicle_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Status"
          }
        },
        "required": [
          "driver_id",
          "service_date",
          "name",
          "active",
          "shift_start_at",
          "shift_end_at",
          "start_lat",
          "start_lng",
          "end_lat",
          "end_lng",
          "vehicle_id",
          "skills",
          "approved_accounts",
          "channel_today",
          "max_hours"
        ],
        "title": "DriverOut",
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "HubOut": {
        "properties": {
          "ac_kw": {
            "title": "Ac Kw",
            "type": "number"
          },
          "ac_points": {
            "title": "Ac Points",
            "type": "integer"
          },
          "dc_kw": {
            "title": "Dc Kw",
            "type": "number"
          },
          "dc_points": {
            "title": "Dc Points",
            "type": "integer"
          },
          "hub_id": {
            "format": "uuid",
            "title": "Hub Id",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "hub_id",
          "name",
          "lat",
          "lng",
          "ac_kw",
          "dc_kw",
          "ac_points",
          "dc_points"
        ],
        "title": "HubOut",
        "type": "object"
      },
...
      "Page_TripRow_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TripRow"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[TripRow]",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 7807 problem details. code is stable and safe to branch on.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "instance"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "TripRow": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "actual_distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Distance M"
          },
          "actual_drop_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Drop At"
          },
          "actual_pickup_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Pickup At"
          },
          "channel": {
            "enum": [
              "blurabbit",
              "ets",
              "uber"
            ],
            "title": "Channel",
            "type": "string"
          },
          "driver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "driver_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Name"
          },
          "drop_address": {
            "title": "Drop Address",
            "type": "string"
          },
          "drop_lat": {
            "title": "Drop Lat",
            "type": "number"
          },
          "drop_lng": {
            "title": "Drop Lng",
            "type": "number"
          },
          "estimate_p50_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimate P50 S"
          },
          "estimate_p80_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimate P80 S"
          },
          "estimate_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimate Source"
          },
          "locked": {
            "default": false,
            "title": "Locked",
            "type": "boolean"
          },
          "luggage": {
            "title": "Luggage",
            "type": "integer"
          },
          "package_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Hours"
          },
          "pax": {
            "title": "Pax",
            "type": "integer"
          },
          "pickup_address": {
            "title": "Pickup Address",
            "type": "string"
          },
          "pickup_lat": {
            "title": "Pickup Lat",
            "type": "number"
          },
          "pickup_lng": {
            "title": "Pickup Lng",
            "type": "number"
          },
          "planned_drop_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Drop At"
          },
          "planned_pickup_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Pickup At"
          },
          "scheduled_pickup_at": {
            "format": "date-time",
            "title": "Scheduled Pickup At",
            "type": "string"
          },
          "service_date": {
            "format": "date",
            "title": "Service Date",
            "type": "string"
          },
          "status": {
            "enum": [
              "booked",
              "assigned",
              "en_route",
              "arrived",
              "started",
              "completed",
              "cancelled",
              "no_show"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "additionalProperties": true,
            "title": "Tags",
            "type": "object"
          },
          "trip_id": {
            "format": "uuid",
            "title": "Trip Id",
            "type": "string"
          },
          "trip_type": {
            "enum": [
              "ets",
              "airport",
              "city",
              "package"
            ],
            "title": "Trip Type",
            "type": "string"
          },
          "unassigned_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unassigned Reason"
          },
          "vehicle_class": {
            "enum": [
              "any",
              "sedan",
              "suv"
            ],
            "title": "Vehicle Class",
            "type": "string"
          }
        },
        "required": [
          "trip_id",
          "service_date",
          "channel",
          "scheduled_pickup_at",
          "pickup_lat",
          "pickup_lng",
          "pickup_address",
          "drop_lat",
          "drop_lng",
          "drop_address",
          "trip_type",
          "package_hours",
          "pax",
          "luggage",
          "vehicle_class",
          "tags",
          "account_id",
          "status",
          "actual_pickup_at",
          "actual_drop_at",
          "actual_distance_m"
        ],
        "title": "TripRow",
        "type": "object"
      },
      "VehicleOut": {
        "properties": {
          "battery_kwh": {
            "title": "Battery Kwh",
            "type": "number"
          },
          "driver_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "driver_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Name"
          },
          "e_roll_wh_per_km": {
            "title": "E Roll Wh Per Km",
            "type": "number"
          },
          "hub_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hub Id"
          },
          "hub_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hub Name"
          },
          "luggage_class": {
            "title": "Luggage Class",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "planning_cap_km": {
            "title": "Planning Cap Km",
            "type": "number"
          },
          "registration": {
            "title": "Registration",
            "type": "string"
          },
          "seats": {
            "title": "Seats",
            "type": "integer"
          },
          "soc_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soc Pct"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "variant": {
            "title": "Variant",
            "type": "string"
          },
          "vehicle_class": {
            "title": "Vehicle Class",
            "type": "string"
          },
          "vehicle_id": {
            "format": "uuid",
            "title": "Vehicle Id",
            "type": "string"
          }
        },
        "required": [
          "vehicle_id",
          "registration",
          "model",
          "variant",
          "vehicle_class",
          "battery_kwh",
          "planning_cap_km",
          "e_roll_wh_per_km",
          "seats",
          "luggage_class",
          "status",
          "hub_id"
        ],
        "title": "VehicleOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Service to service key",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "APIKeyQuery": {
        "description": "JWT for SSE and map tiles",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "description": "JWT from POST /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BluRabbit Dispatch API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/drivers": {
      "get": {
        "operationId": "list_drivers",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriverOut"
                  },
                  "title": "Response List Drivers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Role not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict with the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "summary": "List Drivers",
        "tags": [
          "drivers"
        ]
      }
    },
...
    "/api/v1/hubs": {
      "get": {
        "operationId": "hubs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HubOut"
                  },
                  "title": "Response Hubs",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Role not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict with the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "summary": "Hubs",
        "tags": [
          "vehicles"
        ]
      }
    },
...
    "/api/v1/trips": {
      "get": {
        "operationId": "list_trips",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 16,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TripRow_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Role not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict with the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "summary": "List Trips",
        "tags": [
          "trips"
        ]
      }
    },
...
    "/api/v1/vehicles": {
      "get": {
        "operationId": "list_vehicles",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VehicleOut"
                  },
                  "title": "Response List Vehicles",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Role not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Conflict with the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Rate limited"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyQuery": []
          }
        ],
        "summary": "List Vehicles",
        "tags": [
          "vehicles"
        ]
      }
    },
...
  }
}